project-dapp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==2.4.1
async-timeout==5.0.1
atomicwrites==1.4.1
attrs==25.3.0
bitarray==3.4.2
black==25.1.0
cbor2==5.6.5
cchecksum==0.2.6
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
cytoolz==1.0.1
dataclassy==0.11.1
eip712==0.2.11
eth-account==0.13.7
eth-brownie==1.21.0
eth-event==1.3.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.2.0
exceptiongroup==1.3.1
execnet==2.1.1
fastapi==0.123.0
frozenlist==1.6.0
h11==0.16.0
hexbytes==1.3.1
httptools==0.7.1
hypothesis==6.27.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lark==1.2.2
lazy-object-proxy==1.11.0
lru-dict==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
packaging==23.2
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==5.9.8
py==1.11.0
py-solc-ast==1.2.10
py-solc-x==1.1.1
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pygments-lexer-solidity==0.7.0
pytest==6.2.5
pytest-forked==1.6.0
pytest-xdist==1.34.0
python-dotenv==0.16.0
pyunormalize==16.0.0
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.25.1
semantic-version==2.10.0
six==1.17.0
sortedcontainers==2.4.0
starlette==0.50.0
toml==0.10.2
tomli==2.3.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.38.0
vvm==0.2.1
vyper==0.4.1
watchfiles==1.1.1
wcwidth==0.2.13
web3==6.11.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
